separate llm_generate module from thread
