+++ llm_generate/__manifest__.py
@@ -21,6 +21,8 @@
         "web.assets_backend": [
             "llm_generate/static/src/components/llm_media_form/llm_media_form.js",
             "llm_generate/static/src/components/llm_media_form/llm_media_form.xml",
+            "llm_generate/static/src/components/llm_chat_composer/llm_chat_composer.js",
+            "llm_generate/static/src/components/llm_chat_composer/llm_chat_composer.xml",
             "llm_generate/static/src/models/composer.js",
             "llm_generate/static/src/models/llm_model.js",
             "llm_generate/static/src/models/llm_chat.js",
