+++ llm_generate/models/llm_thread.py
@@ -93,7 +93,7 @@ class LLMThread(models.Model):
             self,
             stream_response,
             LLM_ASSISTANT_SUBTYPE_XMLID,
-            placeholder_text=f"Generated media for: {user_message_body}",
+            placeholder_text=f"<strong>Generated media for:</strong>\n{user_message_body}",
         )
         return assistant_msg
