+++ llm_generate/static/src/components/llm_chat_composer/llm_chat_composer.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<templates xml:space="preserve">
+    <t t-name="llm_generate.LLMChatComposer"
+       t-inherit="llm_thread.LLMChatComposer"
+       t-inherit-mode="extension">
+        
+        <!-- Add the Media Form container below the composer bar -->
+        <xpath expr="//div[hasclass('o_LLMChatComposer_container')]" position="replace">
+            <div class="o_LLMChatComposer_container d-flex align-items-end gap-3 mx-auto" style="max-width: 1200px;">
+                <!-- Standard Chat Composer -->
+                <t t-if="!isMediaGenerationModel">
+                    <LLMChatComposerTextInput record="composerView" />
+                    <t t-if="!isStreaming">
+                        <button
+                            class="o_Composer_actionButton o_Composer_button o_Composer_buttonSend btn btn-primary o-last o-composer-is-compact border-start-0"
+                            t-on-click="_onClickSend"
+                            t-att-disabled="isDisabled"
+                        >
+                            <i class="fa fa-paper-plane-o" />
+                        </button>
+                    </t>
+                    <t t-if="isStreaming">
+                        <button
+                            class="o_Composer_actionButton o_Composer_button o_Composer_buttonStop btn btn-danger o-last o-composer-is-compact border-start-0"
+                            t-on-click="_onClickStop"
+                        >
+                            <i class="fa fa-stop" />
+                        </button>
+                    </t>
+                </t>
+
+                <!-- Media Generation Composer -->
+                <t t-if="isMediaGenerationModel">
+                    <button class="btn btn-secondary w-100" t-on-click="toggleMediaForm">
+                        <i class="fa fa-paint-brush me-1" /> Create Media...
+                    </button>
+                </t>
+            </div>
+
+            <!-- Media Form (conditionally rendered below the composer bar) -->
+            <t t-if="isMediaGenerationModel and state.isMediaFormVisible">
+                <div class="o_LLMChatComposer_mediaFormContainer mx-auto mt-2" style="max-width: 1200px;">
+                    <!-- Pass the llmModel from the thread to the LLMMediaForm -->
+                    <LLMMediaForm model="thread" />
+                </div>
+            </t>
+        </xpath>
+    </t>
+</templates>
